requirements.txt
discord.py==2.5.2
azure-identity==1.25.0
azure-mgmt-compute==37.0.1
python-dotenv==1.1.1
mcstatus==11.1.1
aiohttp>=3.9.0
